requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python==4.8.1.78
ultralytics==8.0.196
numpy==1.24.3
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
requests==2.31.0
websockets==11.0.3
asyncio==3.4.3
threading==1.0
queue==0.1.0
datetime==5.3
json==2.0.9
base64==1.0.0
io==1.0.0
